packages/rules_engine/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "firecore-rules"
version = "0.1.0"
description = "FireCore rules engine — zoning & code rules as data"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.6",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.3",
    "mypy>=1.8",
]

[tool.setuptools.packages.find]
include = ["firecore_rules*"]

[tool.setuptools.package-data]
firecore_rules = ["jurisdictions/*.yaml"]
